Changed the type of attributes to Float rather than Integer

+++ HealthyGrocery/app/models/item.rb
@@ -3,67 +3,67 @@ class Item
   include Mongoid::Paperclip
 
   field :price , type: Float 
-  field :amount , type: Integer
+  field :amount , type: Float
  	field :name , type: String  
  	field :description , type: String  
-	field :rating , type: Integer 
+	field :rating , type: Float 
 	field :status , type: Boolean
 	field :category , type: String
-  field :protein , type: Integer 
-  field :carbohydrate , type: Integer 
-  field :calcium , type: Integer 
-  field :fat , type: Integer 
+  field :protein , type: Float 
+  field :carbohydrate , type: Float 
+  field :calcium , type: Float 
+  field :fat , type: Float 
   field :paused , type: Boolean
   # Author: Hazem
   #Component: 5
   # Attribute: paused, to keep track whether each and every item is paused or not (boolean)
-  field :fiber , type: Integer
-  field :vitamin_a , type: Integer
-  field :vitamin_b , type: Integer
-  field :vitamin_c , type: Integer
-  field :vitamin_d , type: Integer
-  field :vitamin_e , type: Integer
-  field :vitamin_k , type: Integer
-  field :fiber , type: Integer
-  field :folate, type: Integer
-  field :thiamin, type: Integer
-  field :riboflavin, type: Integer
-  field :pantothenic_acid, type: Integer
-  field :choline, type: Integer
-  field :niacin, type: Integer
-  field :betaine, type: Integer
-  field :calcium, type: Integer
-  field :copper, type: Integer
-  field :fluoride, type: Integer
-  field :iron, type: Integer
-  field :magnesium, type: Integer
-  field :manganese, type: Integer
-  field :phosphorus, type: Integer
-  field :selenium, type: Integer
-  field :zinc, type: Integer
-  field :potassium, type: Integer
-  field :sodium, type: Integer
-  field :starch, type: Integer
-  field :sugar, type: Integer
-  field :fatty_acids, type: Integer
-  field :trans_fat, type: Integer
-  field :saturated_fat, type: Integer
-  field :monounsaturated, type: Integer
-  field :polyunsaturated, type: Integer
-  field :trans_monoenoic, type: Integer
-  field :trans_polyenoic, type: Integer
-  field :omega_3, type: Integer
-  field :cartenoids, type: Integer
-  field :tocopherol, type: Integer
-  field :retinol, type: Integer
-  field :beta_carotene, type: Integer
-  field :alpha_carotene, type: Integer
-  field :sucrose, type: Integer
-  field :glucose, type: Integer
-  field :fructose, type: Integer
-  field :maltose, type: Integer
-  field :calories, type: Integer
-  field :minerals, type: Integer
+  field :fiber , type: Float
+  field :vitamin_a , type: Float
+  field :vitamin_b , type: Float
+  field :vitamin_c , type: Float
+  field :vitamin_d , type: Float
+  field :vitamin_e , type: Float
+  field :vitamin_k , type: Float
+  field :fiber , type: Float
+  field :folate, type: Float
+  field :thiamin, type: Float
+  field :riboflavin, type: Float
+  field :pantothenic_acid, type: Float
+  field :choline, type: Float
+  field :niacin, type: Float
+  field :betaine, type: Float
+  field :calcium, type: Float
+  field :copper, type: Float
+  field :fluoride, type: Float
+  field :iron, type: Float
+  field :magnesium, type: Float
+  field :manganese, type: Float
+  field :phosphorus, type: Float
+  field :selenium, type: Float
+  field :zinc, type: Float
+  field :potassium, type: Float
+  field :sodium, type: Float
+  field :starch, type: Float
+  field :sugar, type: Float
+  field :fatty_acids, type: Float
+  field :trans_fat, type: Float
+  field :saturated_fat, type: Float
+  field :monounsaturated, type: Float
+  field :polyunsaturated, type: Float
+  field :trans_monoenoic, type: Float
+  field :trans_polyenoic, type: Float
+  field :omega_3, type: Float
+  field :cartenoids, type: Float
+  field :tocopherol, type: Float
+  field :retinol, type: Float
+  field :beta_carotene, type: Float
+  field :alpha_carotene, type: Float
+  field :sucrose, type: Float
+  field :glucose, type: Float
+  field :fructose, type: Float
+  field :maltose, type: Float
+  field :calories, type: Float
+  field :minerals, type: Float
   #Author: Hazem Amin
   #Component: 5
   #Adding additional attributes for the various items
